Add /api/1/tags/ to edit available compose tags.
I think it's better to have the tag management as normal API
with only `admins` being able to use it. It is easier to
maintain the list of tags like this in comparison to defining
them in the configuration file and ensuring it is mapped
into database properly.

Merges: https://pagure.io/cts/pull-request/6
Signed-off-by: Jan Kaluza <jkaluza@redhat.com>