Preserve repeated query params in paging
The request_args used in pagination are a
werkzeug.datastructures.MultiDict object. A key can be present in there
multiple times. Converting it to a dict results in one value being kept
and all others lost.

This patch instead turns multiple keys into a single key with a list of
values. The `url_for` handles this correctly and produces URL with all
parameters.

JIRA: RHELCMP-12476