Add basic OpenTelemetry integration
By default nothing has changed.

The OTEL_EXPORTER_OTLP_ENDPOINT env variable is driving the telemetry.
When set to "console", the traces are sent to stdout. When set to
anything else, the value is used to configure the exporter.

This commit creates a single span for each request. There is no other
information.

The traceparent value is included in the messages emitted to the message
bus.

JIRA: RHELCMP-11779